moved DetailsSaga and DetailsSlice to MovieDetails
